Converts confirmation token validator to callable method
Using a Laravel validation extension for the token does not make sense with
validation inside models. So the code is moved to the user repository and called
from the controller to ensure a valid token is provided when required. The
validation method can also be considered to reside directly in the controller.