Use an in memory user for the auth token factory
This has the side effect that there is actually no user in the application by
that user id that the token encoder will have access to, so no used data will
be encoded into the factory mocked token. We'll have to figure out another
solution to get a mocked used into the generator.