Injects existing records to faker to prevent unique problems
Faker is not aware of existing records already in the database. With user
records already seeded to the DB when doing the tests, there are a chance that
temporarily records being created during the tests will conflict with the
existing records and throw an error.

To prevent that we inject the current database records into the unique array
in faker to make it become aware of already existing user records and by that
not create any duplicate records.