Ensures unique username from social signup
The name returned from a social provider is not guaranteed to be unique for
Spira. To ensure a seamless process we'll make the username unique for the user
if it is not already.

Initially attempt to make pretty modifications to make it unique, which should
work in most cases. And then fall back on numbering if none of the pretty
methods worked.