Sets a Database placeholder entry
Vanilla does a sanity check before each attempt to save the
configuration. If there is no ‘Database’ key existing in the array to
be saved (in memory config keys are stripped at this time) it will exit
the save method.

See:
https://github.com/vanilla/vanilla/blob/71dc135745ea65c66444e67468e6a467
0421bdac/library/core/class.configuration.php#L1388

Keeping a placeholder Database entry in the config array to be saved
will circumvent this and allow saveConfig to work without saving the DB
credentials.